+++ libs/data-mapper-v2/src/components/functionList/styles.tsx
@@ -1,6 +1,16 @@
-import { makeStyles } from '@fluentui/react-components';
+import { makeStyles, shorthands, tokens, typographyStyles } from '@fluentui/react-components';
+
+const fnIconSize = '12px';
 
 export const useStyles = makeStyles({
+  headerText: {
+    ...typographyStyles.caption1,
+    ...shorthands.borderRadius(tokens.borderRadiusMedium),
+    paddingLeft: tokens.spacingHorizontalXS,
+    fontSize: '13px',
+    marginTop: '8px',
+    marginBottom: '8px',
+  },
   functionSearchBox: {
     width: '210px',
   },
@@ -12,13 +22,38 @@ export const useStyles = makeStyles({
     backgroundColor: '#E8F3FE',
     paddingLeft: '10px',
   },
+  listButton: {
+    height: '30px',
+    width: '100%',
+    display: 'flex',
+    backgroundColor: '#E8F3FE',
+    ...shorthands.border('0px'),
+    ...shorthands.padding('1px 4px 1px 4px'),
+    ':hover': {
+      backgroundColor: '#E8F3FE',
+    },
+  },
+  iconContainer: {
+    height: fnIconSize,
+    flexShrink: '0 !important',
+    flexBasis: fnIconSize,
+    ...shorthands.borderRadius(tokens.borderRadiusCircular),
+    color: tokens.colorNeutralBackground1,
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  functionNameText: {
+    width: '210px',
+    paddingLeft: '4px',
+    paddingRight: '4px',
+    fontSize: '13px',
+    color: '#242424',
+    ...shorthands.overflow('hidden'),
+  },
+  treeItem: {
+    ':hover': {
+      backgroundColor: '#E8F3FE',
+    },
+  },
 });
-
-export const FunctionsSVG = () => (
-  <svg width="40" height="40" viewBox="0 0 32 32" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M14.9149 9.54138C15.2056 9.50052 15.3879 9.50767 15.5277 9.53563C15.6664 9.56338 15.8096 9.62076 16.0139 9.74331C16.3691 9.95643 16.8298 9.84125 17.0429 9.48607C17.256 9.13088 17.1408 8.67019 16.7856 8.45708C16.4899 8.27963 16.1831 8.13701 15.8219 8.06476C15.4631 7.99301 15.0978 8.00001 14.6937 8.05773L14.6765 8.0604C13.8759 8.19383 13.2486 8.607 12.8212 9.2224C12.4076 9.81807 12.2077 10.5624 12.1517 11.3468C12.1504 11.3645 12.1498 11.3824 12.1498 11.4002C12.1498 11.4002 12.1498 11.3995 12.1498 11.4002L12.1496 11.408L12.149 11.4325C12.1483 11.4534 12.1473 11.4809 12.1459 11.5151C12.1431 11.5835 12.1388 11.6743 12.1333 11.7854C12.1223 12.0074 12.1064 12.3053 12.0869 12.6571L12.0817 12.7502H10.8008C10.3866 12.7502 10.0508 13.086 10.0508 13.5002C10.0508 13.9144 10.3866 14.2502 10.8008 14.2502H11.9958C11.9774 14.5642 11.9582 14.8894 11.9386 15.2181C11.8261 17.1056 11.7017 19.0957 11.6522 19.7403C11.5451 21.0244 11.432 21.8778 10.6249 22.3506C10.2223 22.579 9.6062 22.5649 8.93518 22.2294C8.5647 22.0441 8.1142 22.1943 7.92895 22.5648C7.74371 22.9353 7.89388 23.3858 8.26436 23.571C9.19183 24.0347 10.3745 24.2213 11.3719 23.6514L11.3777 23.648C12.9114 22.7533 13.0465 21.0972 13.1373 19.9822L13.1472 19.8625L13.1476 19.8577C13.2067 19.0892 13.3731 16.3974 13.4983 14.2502H15.3008C15.715 14.2502 16.0508 13.9144 16.0508 13.5002C16.0508 13.086 15.715 12.7502 15.3008 12.7502H13.5838C13.6207 12.082 13.6455 11.594 13.6493 11.4344C13.695 10.8284 13.8442 10.3791 14.0533 10.078C14.2496 9.79535 14.5199 9.60913 14.9149 9.54138ZM22.7803 16.2805C23.0732 15.9876 23.0732 15.5127 22.7803 15.2199C22.4874 14.927 22.0126 14.927 21.7197 15.2199L19.1412 17.7983C18.7621 17.1458 18.4841 16.641 18.4208 16.5145C18.4089 16.4905 18.3956 16.4672 18.3811 16.4447C18.3591 16.4104 18.3346 16.3711 18.3077 16.328C18.1551 16.0832 17.9269 15.717 17.6499 15.4609C17.469 15.2935 17.2263 15.129 16.9138 15.0582C16.592 14.9853 16.2672 15.0278 15.9569 15.1595C15.9074 15.1805 15.8604 15.2068 15.8166 15.2378L15.7842 15.2606C15.6982 15.3209 15.6015 15.3888 15.5011 15.4928C15.4557 15.54 15.4105 15.593 15.369 15.6417L15.3466 15.6681C15.3048 15.7173 15.2534 15.7778 15.1833 15.8586C14.912 16.1716 14.9458 16.6453 15.2587 16.9166C15.5717 17.1879 16.0454 17.1541 16.3167 16.8412C16.3919 16.7544 16.4481 16.6883 16.4901 16.6389L16.5113 16.6139C16.5529 16.5651 16.5693 16.5465 16.5776 16.5375L16.5902 16.5282C16.5996 16.5347 16.6133 16.5453 16.6314 16.562C16.6877 16.6141 16.7561 16.6972 16.8433 16.8251C16.9023 16.9116 16.9531 16.9924 17.0105 17.0839C17.0378 17.1274 17.0667 17.1734 17.0986 17.2234C17.2203 17.4587 17.589 18.1209 18.0448 18.8948L15.2197 21.7199C14.9268 22.0127 14.9268 22.4876 15.2197 22.7805C15.5126 23.0734 15.9874 23.0734 16.2803 22.7805L18.9418 20.119C19.3555 20.8277 19.6635 21.3861 19.7305 21.5201C19.739 21.5373 19.7483 21.5541 19.7582 21.5706C19.7731 21.5955 19.7901 21.6245 19.8089 21.6568C19.9383 21.8784 20.1577 22.2542 20.448 22.5292C20.6288 22.7005 20.8705 22.8692 21.1811 22.9507C21.5011 23.0348 21.8349 23.0106 22.1646 22.887C22.1892 22.8777 22.2132 22.8673 22.2367 22.8555C22.5927 22.6775 22.8134 22.4475 23.0408 22.0766C23.2572 21.7234 23.1463 21.2617 22.7932 21.0452C22.44 20.8288 21.9783 20.9396 21.7618 21.2928C21.691 21.4083 21.6526 21.4518 21.6349 21.4692C21.6282 21.4759 21.6205 21.4831 21.6026 21.4939C21.5854 21.4985 21.5746 21.4997 21.5697 21.5L21.5645 21.5002L21.563 21.5001C21.5628 21.5 21.5557 21.4983 21.5425 21.4903C21.5285 21.4818 21.5074 21.4666 21.4796 21.4402C21.4209 21.3846 21.3544 21.3011 21.277 21.1829C21.2203 21.0965 21.1726 21.0156 21.1191 20.9247C21.0998 20.892 21.0797 20.8579 21.0582 20.8218C20.9367 20.5854 20.5321 19.8587 20.0377 19.0232L22.7803 16.2805ZM16.5996 16.5228C16.5996 16.5228 16.6 16.5227 16.5996 16.5228V16.5228ZM16.5976 16.5228C16.599 16.5227 16.5997 16.5227 16.5996 16.5228L16.5976 16.5228Z"
-      fill="#0F6CBD"
-    />
-  </svg>
-);
