+++ libs/data-mapper-v2/src/images/FunctionIcons/DataType16Icons.tsx
