+++ libs/data-mapper-v2/src/images/FunctionIcons/DataType24Icons.tsx
