+++ libs/data-mapper-v2/src/images/FunctionIcons/FunctionIcons.tsx
